+++ src/common/test/SyncRequestTest.cc
@@ -52,7 +52,7 @@ TEST(SyncRequestTest, NoThread)
   {
     TimeBuffer_p buffer = std::make_shared<TimeBuffer>(sensors, time);
     buffer->syncRequest(sqla);
-    EXPECT_EQ(2*24 + 1, buffer->size());
+    EXPECT_EQ(size_t(2*24 + 1), buffer->size());
   }
 }
 
@@ -67,7 +67,7 @@ TEST(SyncRequestTest, Thread)
   {
     IndexBuffer_p buffer = std::make_shared<IndexBuffer>(3600, sensors, time);
     buffer->syncRequest(sqla);
-    EXPECT_EQ(2*24 + 1, buffer->size());
+    EXPECT_EQ(size_t(2*24 + 1), buffer->size());
   }
 }
 
@@ -83,15 +83,15 @@ TEST(SyncRequestTest, Cached)
   { const TimeSpan time(s2t("2013-04-01 00:00:00"), s2t("2013-04-03 00:00:00"));
     TimeBuffer_p buffer = std::make_shared<TimeBuffer>(sensors, time);
     buffer->syncRequest(ca);
-    EXPECT_EQ(1, sqla->countPost());
-    EXPECT_EQ(2*24 + 1, buffer->size());
+    EXPECT_EQ(1u, sqla->countPost());
+    EXPECT_EQ(size_t(2*24 + 1), buffer->size());
   }
 
   { const TimeSpan time(s2t("2013-04-01 00:00:00"), s2t("2013-04-02 00:00:00"));
     TimeBuffer_p buffer = std::make_shared<TimeBuffer>(sensors, time);
     buffer->syncRequest(ca);
-    EXPECT_EQ(1, sqla->countPost());
-    EXPECT_EQ(1*24 + 1, buffer->size());
+    EXPECT_EQ(1u, sqla->countPost());
+    EXPECT_EQ(size_t(1*24 + 1), buffer->size());
   }
 }
 
@@ -108,21 +108,21 @@ TEST(SyncRequestTest, CachedMulti)
   { const TimeSpan time(s2t("2014-03-01 00:00:00"), s2t("2014-03-01 06:00:00"));
     TimeBuffer_p buffer = std::make_shared<TimeBuffer>(make_set<Sensor_s>(sensor1), time);
     buffer->syncRequest(ca);
-    EXPECT_EQ(1, sqla->countPost());
-    EXPECT_EQ(7, buffer->size());
+    EXPECT_EQ(1u, sqla->countPost());
+    EXPECT_EQ(7u, buffer->size());
   }
 
   { const TimeSpan time(s2t("2014-03-01 03:00:00"), s2t("2014-03-01 09:00:00"));
     TimeBuffer_p buffer = std::make_shared<TimeBuffer>(make_set<Sensor_s>(sensor2), time);
     buffer->syncRequest(ca);
-    EXPECT_EQ(2, sqla->countPost());
-    EXPECT_EQ(7, buffer->size());
+    EXPECT_EQ(2u, sqla->countPost());
+    EXPECT_EQ(7u, buffer->size());
   }
 
   { const TimeSpan time(s2t("2014-03-01 00:00:00"), s2t("2014-03-01 09:00:00"));
     TimeBuffer_p buffer = std::make_shared<TimeBuffer>(sensors, time);
     buffer->syncRequest(ca);
-    EXPECT_EQ(4, sqla->countPost());
-    EXPECT_EQ(20, buffer->size());
+    EXPECT_EQ(4u, sqla->countPost());
+    EXPECT_EQ(20u, buffer->size());
   }
 }
